backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Киприн Вячеслав",email = "kiprinvs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (==5.2.1)",
    "djangorestframework (==3.16.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-phonenumber-field[phonenumberslite] (>=8.1.0,<9.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework-simplejwt (==5.3.1)",
    "djoser (>=2.3.1,<3.0.0) ; python_version >= \"3.11\" and python_version < \"4.0\"",
    "gunicorn (>=23.0.0,<24.0.0)",
    "social-auth-app-django (>=5.5.1,<6.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-django = "^4.11.1"
ruff = "^0.12.1"
psycopg2 = "^2.9.10"
pytest-cov = "^6.2.1"
